Finished converting the PreLoadingScreen methods into common operations. Changed the NetworkUtils into NetworkUtilExtensions to comply with the existing NetworkUtils from LaminariaCore-General
